移动浏览器版本的 AV._ajax 到 ajax-browser.js
