+++ lib/av.js
@@ -34,10 +34,7 @@ require('./status')(AV);
 require('./search')(AV);
 require('./insight')(AV);
 
-// Node.js platform specific patches
-if (AV._isNode) {
-  AV._ajax = require('./browserify-wrapper/ajax');
-}
+AV._ajax = require('./browserify-wrapper/ajax');
 
 // Backward compatibility
 AV.AV = AV;
