+++ lib/utils.js
@@ -246,96 +246,6 @@ module.exports = function(AV) {
     return new Date(Date.UTC(year, month, day, hour, minute, second, milli));
   };
 
-  AV._ajaxIE8 = function(method, url, data) {
-    var promise = new AV.Promise();
-    var xdr = new XDomainRequest();
-    xdr.onload = function() {
-      var response;
-      try {
-        response = JSON.parse(xdr.responseText);
-      } catch (e) {
-        promise.reject(e);
-      }
-      if (response) {
-        promise.resolve(response);
-      }
-    };
-    xdr.onerror = xdr.ontimeout = function() {
-      // Let's fake a real error message.
-      var fakeResponse = {
-        responseText: JSON.stringify({
-          code: AV.Error.X_DOMAIN_REQUEST,
-          error: "IE's XDomainRequest does not supply error info."
-        })
-      };
-      promise.reject(xdr);
-    };
-    xdr.onprogress = function() {};
-    xdr.open(method, url);
-    xdr.send(data);
-    return promise;
-  };
-
-   AV._useXDomainRequest = function() {
-       if (typeof(XDomainRequest) !== "undefined") {
-           // We're in IE 8+.
-           if ('withCredentials' in new XMLHttpRequest()) {
-               // We're in IE 10+.
-               return false;
-           }
-           return true;
-       }
-       return false;
-   };
-
-  AV._ajax = function(method, url, data, success, error) {
-    var options = {
-      success: success,
-      error: error
-    };
-
-    if (AV._useXDomainRequest()) {
-      return AV._ajaxIE8(method, url, data)._thenRunCallbacks(options);
-    }
-
-    var promise = new AV.Promise();
-    var handled = false;
-
-    var xhr = new AV.XMLHttpRequest();
-    xhr.onreadystatechange = function() {
-      if (xhr.readyState === 4) {
-        if (handled) {
-          return;
-        }
-        handled = true;
-
-        if (xhr.status >= 200 && xhr.status < 300) {
-          var response;
-          try {
-            response = JSON.parse(xhr.responseText);
-          } catch (e) {
-            promise.reject(e);
-          }
-          if (response) {
-            promise.resolve(response, xhr.status, xhr);
-          }
-        } else {
-          promise.reject(xhr);
-        }
-      }
-    };
-    xhr.open(method, url, true);
-    xhr.setRequestHeader("Content-Type", "text/plain");  // avoid pre-flight.
-    if (AV._isNode) {
-      // Add a special user agent just for request from node.js.
-      xhr.setRequestHeader("User-Agent",
-                           "AV/" + AV.VERSION +
-                           " (NodeJS " + process.versions.node + ")");
-    }
-    xhr.send(data);
-    return promise._thenRunCallbacks(options);
-  };
-
   // A self-propagating extend function.
   AV._extend = function(protoProps, classProps) {
     var child = inherits(this, protoProps, classProps);
